+++ covidmetrics/application/staff/forms.py
@@ -1,6 +1,6 @@
 """Sign-up & log-in forms."""
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, BooleanField, HiddenField, SelectField, RadioField, IntegerField
+from wtforms import DecimalField, PasswordField, SubmitField, BooleanField, HiddenField, SelectField, RadioField, IntegerField
 from wtforms.validators import DataRequired, Email, EqualTo, Length, Optional, NumberRange
 
 
@@ -11,13 +11,13 @@ class StaffUpdateForm(FlaskForm):
 
     facility_id = SelectField(u'Facility', validators=[DataRequired()])
 
-    required = IntegerField(
+    required = DecimalField(
         'Required',
         validators=[
             DataRequired()
         ]
     )
-    available = IntegerField(
+    available = DecimalField(
         'Available',
         validators=[
             DataRequired()
